requirements.txt

# Backend API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Deep Learning & Model Training (FastAI)
fastai>=2.7.12
torch>=2.6.0
torchvision>=0.17.0
timm>=0.9.7
ipython>=8.0.0
jupyter>=1.0.0
scipy>=1.10.0

# Data Processing
numpy>=1.26.2
pandas>=2.1.3
Pillow>=11.0.0
scikit-learn>=1.3.2
openpyxl>=3.1.2

# Visualization (optional, for training)
matplotlib>=3.8.2
seaborn>=0.13.0

# Utilities
requests>=2.31.0
reportlab>=3.6.14


